AdVitam(Sleepiness label)/requirements.txt
# Core data processing and analysis
numpy>=1.24.0
pandas>=2.2.3
scipy>=1.11.0

# Machine learning
scikit-learn>=1.6.1
torch>=2.7.0
xgboost>=3.0.2

# Physiological signal processing
bioread>=2025.5.2
neurokit2>=0.2.0

# Visualization
matplotlib>=3.10.3
seaborn>=0.13.2

# Jupyter and notebooks
jupyter>=1.1.1
ipykernel>=6.29.5
nb-clean>=4.0.1

# Configuration and utilities
pyyaml>=6.0
openpyxl>=3.1.5
pyarrow>=20.0.0

# Development and testing (optional)
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0

# dotenv>=1.0.1